requirements.txt
fastapi>=0.110.0
selenium>=4.18.0
webdriver-manager>=4.0.1
torch==2.1.2
uvicorn>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.1
deepl>=1.16.0
spotipy>=2.23.0
lyricsgenius>=3.0.0
pandas>=2.2.0
transformers>=4.40.0

